backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
fastapi==0.110.1
uvicorn==0.30.1
google-api-python-client==2.129.0
google-auth-oauthlib==1.2.0
google-auth==2.31.0
google-generativeai==0.4.1
transformers==4.41.2
openai==1.35.3
torch==2.3.1
python-dotenv==1.0.1
prometheus-fastapi-instrumentator==6.1.0
prometheus-client==0.20.0
sqlmodel==0.0.16
psycopg2-binary==2.9.9
alembic==1.13.2
pytest==8.2.2
httpx==0.27.0
ruff==0.4.8
apscheduler==3.10.4
email-validator==2.1.0